# backend/openapi/openapi.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GCM Backend API",
    "version": "1.0.0",
    "description": "Versioned public API contract for GCM v2. All new clients should use /api/v1/* endpoints."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin backend"
    }
  ],
...
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" }
        }
      }
    },
...
    "/api/v1/auth/sse-token": {
      "get": {
        "tags": ["Auth"],
        "summary": "Issue short-lived SSE token",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "SSE token issued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SseTokenResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Users list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create or update user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpsert" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upsert success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpsertSuccess" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Delete success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteSuccess" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" }
        }
      }
    },
    "/api/v1/companies": {
      "get": {
        "tags": ["Companies"],
        "summary": "List companies",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Companies list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Company" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Companies"],
        "summary": "Create or update company",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompanyUpsert" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upsert success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpsertSuccess" }
              }
            }
          },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/api/v1/companies/{id}": {
      "delete": {
        "tags": ["Companies"],
        "summary": "Delete company",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Delete success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteSuccess" }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List projects",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Projects list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Project" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create or update project",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectUpsert" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upsert success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpsertSuccess" }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "delete": {
        "tags": ["Projects"],
        "summary": "Delete project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Delete success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteSuccess" }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips": {
      "get": {
        "tags": ["Trips"],
        "summary": "List trips",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Trips list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Trip" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Trips"],
        "summary": "Create or update trip",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TripUpsert" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upsert success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpsertSuccess" }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{id}": {
      "delete": {
        "tags": ["Trips"],
        "summary": "Delete trip",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Delete success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteSuccess" }
              }
            }
          }
        }
      }
    },
    "/api/v1/services": {
      "get": {
        "tags": ["Services"],
        "summary": "List services",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Services list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Service" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Services"],
        "summary": "Create or update service",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ServiceUpsert" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upsert success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpsertSuccess" }
              }
            }
          }
        }
      }
    },
    "/api/v1/services/{id}": {
      "delete": {
        "tags": ["Services"],
        "summary": "Delete service",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Delete success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteSuccess" }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities": {
      "get": {
        "tags": ["Facilities"],
        "summary": "List facilities",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Facilities list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Facility" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Facilities"],
        "summary": "Create or update facility",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FacilityUpsert" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upsert success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpsertSuccess" }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{id}": {
      "delete": {
        "tags": ["Facilities"],
        "summary": "Delete facility",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Delete success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteSuccess" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/drivers/{id}": {
      "delete": {
        "tags": ["Drivers"],
        "summary": "Delete driver",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Delete success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteSuccess" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "description": "Authenticated user profile. Session token is issued via HttpOnly cookie, not response body.",
        "properties": {
          "user": { "$ref": "#/components/schemas/User" }
        },
        "additionalProperties": true
      },
      "SseTokenResponse": {
        "type": "object",
        "required": ["token", "expiresInSeconds"],
        "properties": {
          "token": { "type": "string" },
          "expiresInSeconds": { "type": "integer" }
        },
        "additionalProperties": true
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string" }
        },
        "additionalProperties": true
      },
...
    }
  }
}